\name{fasta_file}
\docType{data}
\alias{fasta_file}
\title{FASTA file (subdirectory "data")}
\description{
The FASTA file (unpacked) in the subdirectory data of the PopGenome package 
have to be stored in a folder (multiple files can be stored in this folder).
This folder is than the input for the \code{readData} function.
}
\keyword{datasets}
