\name{get.status-methods}
\docType{methods}
\alias{get.status-methods}
\alias{get.status,GENOME-method}



\title{State of Calculations}

\description{
Some informations about the definitions of poulations and subsites.
}

\usage{

\S4method{get.status}{GENOME}(object)

}

 \arguments{

  \item{object}{An object of class \code{"GENOME"}}
  
}


\examples{
# get.status(GENOME.class)
}

\keyword{methods}

