\name{gff_file}
\docType{data}
\alias{gff_file}
\title{GFF file (subdirectory "data")}
\description{
A typical GFF file which should be stored in a folder (for example in "GFF").
This folder is the input for the \code{readData(...,gffpath="GFF")} function.
The corresponding FASTA file is stored in the "data" subdirectory 
of the PopGenome package. It has to be stored in a folder with the 
SAME NAME as the GFF file (for example in "FASTA").
\code{readData("FASTA", gffpath="GFF")}

}
\keyword{datasets}
