\name{load.session}

\docType{methods}
\alias{load.session}


\title{Loading a PopGenome-session}


\description{
This function loads a PopGenome-session from the current workspace.
Saving with save().
}

\usage{

load.session()

}

\arguments{
no arguments !
}  

\details{
coming soon !
}

\value{
coming soon !
}


\examples{

# GENOME.class <- readData("...\Alignments")
# save.session(GENOME.class)
# q()
# R
# library(PopGenome)
# GENOME.class <-load.session()

}



\keyword{methods}
