\name{show.slots-methods}
\docType{methods}
\alias{show.slots-methods}
\alias{show.slots,GENOME-method}
\title{Show Slots of class GENOME}
\description{
coming soon ...
}
\section{Methods}{
\describe{

\item{object = "GENOME"}{ coming soon ... }
}}

\examples{
# show.slots(GENOME.class)
}

\keyword{methods}


