\name{snp_file}
\docType{data}
\alias{snp_file}
\title{SNP file (variant call (1001 Genome project)}
\description{
A .SNP file stored in the directory "data" of the PopGenome package.
The file contains variant calls for exactly one individual.
Put all files (individuals of interest) in one folder (for example "SNP").
\code{readSNP("SNP",CHR=1)}
}
\keyword{datasets}
