\name{vcf_file}
\docType{data}
\alias{vcf_file}
\title{VCF file (subdirectory "data")}
\description{
A VCF file stored in the directory "data" of the PopGenome package.
The file (unpacked) have to be stored in a folder (for example "VCF").
Note, many VCF-files can be stored in this folder. If the 
VCF file is too big, split it into chunks (by position) !
PopGenome is able to concatenate them afterwards.
\code{readData("VCF",format="VCF",FAST=TRUE)}
}
\keyword{datasets}
