\name{fasta_file}
\docType{data}
\alias{fasta_file}
\title{FASTA file (subdirectory "data")}
\description{
The FASTA files (unpacked) in the subdirectory "data" of the PopGenome package 
have to be stored in a folder (multiple files can be stored in this folder).
The folder name is then used as the input for the \code{readData} function.
}
\keyword{datasets}
