\name{get.individuals-methods}
\docType{methods}

\alias{get.individuals,GENOME-method}
\alias{get.individuals-methods}


\title{Print the names/IDs of individuals}


\description{
Extract the names/IDs of individuals.
}

\usage{

\S4method{get.individuals}{GENOME}(object,region=FALSE)

}

\arguments{

  \item{object}{an object of class \code{"GENOME"}}
  \item{region}{a vector of regions. Default: ALL}
}



\examples{

# GENOME.class <- readData("\home\Alignments")
# get.individuals(GENOME.class)

}



\keyword{methods}

