\name{get.status-methods}
\docType{methods}
\alias{get.status-methods}
\alias{get.status,GENOME-method}
\title{State of Calculations}

\description{
coming soon ...
}
\section{Methods}{
\describe{

\item{object = "GENOME"}{ coming soon ...}
}}

\examples{
# get.status(GENOME.class)
}

\keyword{methods}

